fix(images): relocated 404 & 500 from samples to images
