+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,5 +1,5 @@
 import { ErrorPage } from '../../components/shared/ErrorPage/ErrorPage';
-import image from '../../images/samples/404_large_chart.png';
+import image from '../../images/404_large_chart.png';
 import { NOT_FOUND_SUBTITLE, NOT_FOUND_TITLE } from '../../utils/constants';
 
 export const NotFoundPage = () => {
